move: Rename test data
