+++ test/test-input-scene-options.ts
