+++ test/test-output-data-scene.ts
